refactor(shadow): colocate API groups with contracts
