+++ shadow/shared/api/health-check.ts
@@ -1,3 +1,4 @@
+import { HttpApiEndpoint, HttpApiGroup } from '@effect/platform'
 import { Schema } from 'effect'
 
 export const HealthCheckResponse = Schema.Struct({
@@ -5,3 +6,7 @@ export const HealthCheckResponse = Schema.Struct({
 })
 
 export type HealthCheckResponse = Schema.Schema.Type<typeof HealthCheckResponse>
+
+export class HealthCheckApi extends HttpApiGroup.make('healthCheck').add(
+  HttpApiEndpoint.get('check', '/health-check').addSuccess(HealthCheckResponse),
+) {}
